added file handlers, party mode

+++ src/com/synchophy/server/scanner/FileScanner.java
@@ -1,22 +1,21 @@
 package com.synchophy.server.scanner;
 
 import java.io.File;
+import java.io.IOException;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import org.farng.mp3.AbstractMP3Tag;
-import org.farng.mp3.MP3File;
-import org.farng.mp3.id3.AbstractID3v2;
-import org.farng.mp3.id3.ID3v1;
-import org.farng.mp3.id3.ID3v1_1;
-
+import com.synchophy.server.ConfigManager;
 import com.synchophy.server.db.DatabaseManager;
+import com.synchophy.server.player.IMediaPlayer;
+import com.synchophy.server.scanner.handler.IFileHandler;
 import com.synchophy.util.StringUtils;
 
 public class FileScanner {
@@ -24,15 +23,58 @@ public class FileScanner {
 	private static final int MAX_BATCH_SIZE = 100;
 
 	private String basePath;
+	private String scanPath;
+	private boolean insertOnly;
 	private PreparedStatement sth;
 	private PreparedStatement errorSth;
 	private int batchSize = 0;
 	private int errorBatchSize = 0;
-	private static final String[] supportedExts = TaggedFile.formats(); 
+	private static final String[] supportedExts = TaggedFile.formats();
+	private static Map handlers = new HashMap();
 
 	public FileScanner(String basePath) {
+		this(basePath, basePath);
+	}
 
+	public FileScanner(String basePath, String scanPath) {
+		this(basePath, scanPath, false);
+	}
+
+	public FileScanner(String basePath, String scanPath, boolean insertOnly) {
 		this.basePath = basePath;
+		this.scanPath = scanPath;
+		this.insertOnly = insertOnly;
+		init();
+	}
+
+	private void init() {
+		List handlers = ConfigManager.getInstance().getFileHandlers();
+		Iterator i = handlers.iterator();
+		while (i.hasNext()) {
+
+			String handlerClass = (String) i.next();
+			try {
+
+				Class clazz = Class.forName(handlerClass);
+				Object obj = clazz.newInstance();
+				if (obj instanceof IFileHandler == false) {
+					throw new RuntimeException(
+							"Invalid file handler specified: " + handlerClass);
+				}
+				IFileHandler fileHandler = (IFileHandler) obj;
+				FileScanner.registerHandler(fileHandler.getExtension(),
+						fileHandler);
+			} catch (Exception e) {
+				throw new RuntimeException("Could not register file handler: "
+						+ handlerClass);
+
+			}
+		}
+
+	}
+
+	public static void registerHandler(String extension, IFileHandler handler) {
+		handlers.put(extension, handler);
 	}
 
 	public void scan() {
@@ -42,76 +84,52 @@ public class FileScanner {
 		sth = DatabaseManager
 				.getInstance()
 				.prepare(
-						"insert into import (file, track, artist, artist_sort, artist_key, album, album_sort, album_key, title, title_sort, title_key,  size) "
-								+ " values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
+						"insert into import (file, track, artist, artist_sort, artist_key, album, album_sort, album_key, title, title_sort, title_key,  size, orchestra, orchestra_sort, orchestra_key) "
+								+ " values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
 		if (sth == null) {
 			return;
 		}
 		errorSth = DatabaseManager.getInstance().prepare(
 				"insert into import_error (file, message) values (?, ?)");
 
-		scan(new File(this.basePath));
+		scan(new File(this.scanPath));
 
 		executeBatch();
 		executeErrorBatch();
-		
-		String deleteSQL = "DELETE FROM song where file not in (select file from IMPORT)";
-		DatabaseManager.getInstance().executeQuery(deleteSQL);
 
-		String insertSQL = "INSERT INTO song (file, track, artist, artist_sort, artist_key, album, album_sort, album_key, title, title_sort, title_key,  size) "
-				+ " (SELECT file, track, artist, artist_sort, artist_key, album, album_sort, album_key, title, title_sort, title_key, size from import where file not in (select file from song))";
+//		AlbumArtScanner scanner = new AlbumArtScanner();
+//		List rows = DatabaseManager
+//				.getInstance()
+//				.query("select artist_sort, album_sort from import group by artist_sort, album_sort",
+//						new Object[0], new String[] { "artist", "album" });
+//		Iterator i = rows.iterator();
+//		while (i.hasNext()) {
+//			Map row = (Map) i.next();
+//			try {
+//				scanner.scan("import", (String) row.get("artist"),
+//						(String) row.get("album"));
+//			} catch (IOException e) {
+//				e.printStackTrace();
+//			}
+//		}
+
+		if (!insertOnly) {
+			String deleteSQL = "DELETE FROM song where file not in (select file from IMPORT)";
+			DatabaseManager.getInstance().executeQuery(deleteSQL);
+		}
+
+		String insertSQL = "INSERT INTO song (file, track, artist, artist_sort, artist_key, album, album_sort, album_key, title, title_sort, title_key,  size, orchestra, orchestra_sort, orchestra_key) "
+				+ " (SELECT file, track, artist, artist_sort, artist_key, album, album_sort, album_key, title, title_sort, title_key, size, orchestra, orchestra_sort, orchestra_key from import where file not in (select file from song))";
 		DatabaseManager.getInstance().executeQuery(insertSQL);
-		// String updateSQL =
-		// "UPDATE song set song.artist = (select artist from import where song.file = import.file), song.artist_sort = (select artist_sort from import where song.file = import.file), song.artist_key = (select artist_key from import where song.file = import.file), "
-		// +
-		// "song.album = (select album from import where song.file = import.file), song.album_sort = (select album_sort from import where song.file = import.file), song.album_key = (select album_key from import where song.file = import.file), "
-		// +
-		// "song.title = (select title from import where song.file = import.file), song.title_sort = (select title_sort from import where song.file = import.file), song.title_key = (select title_key from import where song.file = import.file), "
-		// +
-		// "song.size = (select size from import where song.file = import.file)";
-		// DatabaseManager.getInstance().executeQuery(updateSQL);
 
-		deleteSQL = "DELETE FROM bad_song where file not in (select file from import_error)";
-		DatabaseManager.getInstance().executeQuery(deleteSQL);
+		if (!insertOnly) {
+			String deleteSQL = "DELETE FROM bad_song where file not in (select file from import_error)";
+			DatabaseManager.getInstance().executeQuery(deleteSQL);
+		}
 
 		insertSQL = "INSERT INTO bad_song (file, message) (select file, message from import_error where file not in (select file from bad_song))";
 		DatabaseManager.getInstance().executeQuery(insertSQL);
 
-		// updateSQL =
-		// "update bad_song set message = (select message from import_error where bad_song.file = import_error.file)";
-		// DatabaseManager.getInstance().executeQuery(updateSQL);
-
-		// String updateSQL = "update song set song.track = vals.track"
-		// +
-		// " song.artist = vals.artist, song.artist_sort = vals.artist_sort, song.artist_key = vals.artist_key, "
-		// +
-		// "song.album = vals.album, song.album_sort = vals.album_sort, song.album_key = vals.album_key, "
-		// +
-		// "song.title = vals.title, song.title_sort = vals.title_sort, song.title_key = vals.album_key, "
-		// + "song.size = vals.size "
-		// DatabaseManager.getInstance()
-		// .executeQuery("merge into song using (select file, track, artist, artist_sort, artist_key, album, album_sort, album_key, title, title_sort, title_key, size from import) as vals(file, track, artist, artist_sort, artist_key, album, album_sort, album_key, title, title_sort, title_key, size)"
-		// + " on song.file = vals.file"
-		// + " when matched then"
-		// + " update set song.track = vals.track, "
-		// +
-		// " song.artist = vals.artist, song.artist_sort = vals.artist_sort, song.artist_key = vals.artist_key, "
-		// +
-		// "song.album = vals.album, song.album_sort = vals.album_sort, song.album_key = vals.album_key, "
-		// +
-		// "song.title = vals.title, song.title_sort = vals.title_sort, song.title_key = vals.album_key, "
-		// + "song.size = vals.size"
-		// +
-		// " when not matched then insert values(null, vals.file, vals.track, vals.artist, vals.artist_sort, vals.artist_key, "
-		// + "vals.album, vals.album_sort, vals.album_key, "
-		// + "vals.title, vals.title_sort, vals.title_key, vals.size)");
-		// DatabaseManager.getInstance()
-		// .executeQuery("merge into bad_song using (select file, message from import_error) as vals(file, message)"
-		// + " on bad_song.file = vals.file"
-		// + " when matched then"
-		// + " update set bad_song.message = vals.message"
-		// +
-		// " when not matched then insert values(null, vals.file, vals.message)");
 		DatabaseManager.getInstance().executeQuery("delete from import");
 		DatabaseManager.getInstance().executeQuery("delete from import_error");
 
@@ -149,25 +167,46 @@ public class FileScanner {
 					scan(file);
 					continue;
 				}
+				if (hasHandler(file)) {
+					handle(file);
+				}
+			}
+		}
+		fileNames = path.list();
+		for (int i = 0; i < fileNames.length; i++) {
+			String fileName = fileNames[i];
+			File file = new File(path, fileName);
+			if (file.exists() && !file.isDirectory()) {
 				insertFile(file);
 			}
 		}
+	}
+
+	private boolean hasHandler(File file) {
+		String ext = StringUtils.getExtension(file);
+		return handlers.containsKey(ext);
+	}
+
+	private void handle(File file) {
+		String ext = StringUtils.getExtension(file);
+
+		IFileHandler handler = (IFileHandler) handlers.get(ext);
+		handler.handle(this, file);
 	}
 
 	private void insertFile(File file) {
 
 		try {
 			System.err.println("Inserting " + file.getAbsolutePath());
-			String filename = file.getName();
-			int lastPeriod = filename.lastIndexOf('.');
-			if (lastPeriod < 0) {
+
+			String ext = StringUtils.getExtension(file);
+			if ("".equals(ext)) {
 				insertFileError(file, "Unsupported file name");
 				return;
 			}
 
-			String ext = filename.substring(lastPeriod + 1).toUpperCase();
 			boolean supportedFile = false;
-			for(int i = 0; i < supportedExts.length; i++) {
+			for (int i = 0; i < supportedExts.length; i++) {
 				if (ext.equals(supportedExts[i]) == true) {
 					supportedFile = true;
 				}
@@ -186,7 +225,7 @@ public class FileScanner {
 			} catch (Exception e) {
 				e.printStackTrace();
 			}
-			insertFileError(file, "No tag found in " + filename);
+			insertFileError(file, "No tag found in " + file.getAbsolutePath());
 		} catch (Exception e) {
 			insertFileError(file, e.getMessage());
 		}
@@ -195,6 +234,8 @@ public class FileScanner {
 	private void insertFile(File file, TaggedFile tfile) {
 
 		try {
+			file = moveFile(file, tfile);
+
 			// :file, :track, :artist, :album, :title, :size
 			sth.setString(1, file.getAbsolutePath());
 			sth.setString(2, StringUtils.cleanTrack(tfile.getTrack()));
@@ -213,6 +254,11 @@ public class FileScanner {
 							+ tfile.getTitle());
 			sth.setString(11, StringUtils.sortLetter(tfile.getTitle()));
 			sth.setLong(12, file.length());
+			sth.setString(13, tfile.getOrchestra());
+			sth.setString(14,
+					StringUtils.alphabetizeLinguistically(tfile.getOrchestra()));
+			sth.setString(15, StringUtils.sortLetter(tfile.getOrchestra()));
+
 			sth.addBatch();
 			batchSize++;
 			if (batchSize > MAX_BATCH_SIZE) {
@@ -221,6 +267,40 @@ public class FileScanner {
 		} catch (SQLException e) {
 			insertFileError(file, e.getMessage());
 		}
+	}
+
+	private File moveFile(File file, TaggedFile tfile) {
+
+		String artist = StringUtils.cleanFilename(tfile.getOrchestra());
+		if ("".equals(artist)) {
+			artist = StringUtils.cleanFilename(tfile.getArtist());
+		}
+		String toPath = basePath + "//" + artist + "/" + tfile.getAlbum() + "/"
+				+ file.getName();
+		toPath = StringUtils.removeDoubleSlashes(toPath);
+
+		if (!toPath.equals(file.getAbsolutePath())) {
+			System.err.println("I am moving " + file.getAbsolutePath() + " to "
+					+ toPath);
+			File toFile = new File(toPath);
+			File parent = toFile.getParentFile();
+			if (!parent.exists()) {
+				parent.mkdirs();
+			}
+			if (parent.exists()) {
+				if (file.renameTo(toFile)) {
+					file.delete();
+					return toFile;
+				} else {
+					System.err.println("Could not delete "
+							+ file.getAbsolutePath());
+				}
+			} else {
+				System.err.println("Could not mkdirs "
+						+ toFile.getParentFile().getAbsolutePath());
+			}
+		}
+		return file;
 	}
 
 	private void insertFileError(File file, String message) {
@@ -262,17 +342,18 @@ public class FileScanner {
 		System.gc();
 	}
 
-	public static void main(String[] args) {
+	public static void main(String[] args) throws Exception {
 		// does not have track info. Attempt filename parsing
-		// /home/tthomas/MUSIC/Clinic/Winchester Cathedral/07 - clinic - winchester_cathedral -.mp3'
+		// /home/tthomas/MUSIC/Clinic/Winchester Cathedral/07 - clinic -
+		// winchester_cathedral -.mp3'
 		Logger.global.setLevel(Level.SEVERE);
 
 		try {
-			String musicPath = System.getProperty("music.path", "/Music");
+			String musicPath = ConfigManager.getInstance().getMusicPath();
 			Date start = new Date();
 			FileScanner scanner = new FileScanner(musicPath);
 			scanner.scan();
-			System.err.println("Scanned in "
+			System.err.println("Scanned music "
 					+ ((new Date().getTime() - start.getTime()) / 1000)
 					+ "secs");
 		} finally {
